[mob] Move model
