+++ mobile/lib/utils/file_uploader.dart
@@ -24,6 +24,7 @@ import 'package:photos/events/files_updated_event.dart';
 import 'package:photos/events/local_photos_updated_event.dart';
 import 'package:photos/events/subscription_purchased_event.dart';
 import 'package:photos/main.dart';
+import "package:photos/models/api/metadata.dart";
 import "package:photos/models/backup/backup_item.dart";
 import "package:photos/models/backup/backup_item_status.dart";
 import 'package:photos/models/file/file.dart';
@@ -34,7 +35,6 @@ import 'package:photos/module/upload/model/upload_url.dart';
 import "package:photos/module/upload/service/multipart.dart";
 import "package:photos/service_locator.dart";
 import 'package:photos/services/collections_service.dart';
-import "package:photos/services/file_magic_service.dart";
 import 'package:photos/services/local_sync_service.dart';
 import "package:photos/services/preview_video_store.dart";
 import 'package:photos/services/sync_service.dart';
@@ -751,7 +751,8 @@ class FileUploader {
       if (SyncService.instance.shouldStopSync()) {
         throw SyncStopRequestedError();
       }
-      final stillLocked = await _uploadLocks.isLocked(lockKey, _processType.toString());
+      final stillLocked =
+          await _uploadLocks.isLocked(lockKey, _processType.toString());
       if (!stillLocked) {
         _logger.warning('file ${file.tag} report paused is missing');
         throw LockFreedError();
